+++ src/api/waitAndSubmitMessageToServer.ts
