+++ src/components/RefreshMessagesView.tsx
@@ -6,7 +6,7 @@ import {
     StyleSheet,
 } from "react-native";
 import { connect } from "react-redux";
-import { requestFetchMessages } from "../state/action-creators/requestFetchPosts";
+import { requestFetchMessages } from "../state/action-creators/requestFetchMessages";
 import { MapStateToProps } from "../state/state/MapStateToProps";
 
 const styles = StyleSheet.create({
