+++ src/state/sagas/submitMessageSaga.ts
@@ -1,5 +1,5 @@
 import { call, put, select, takeEvery } from "redux-saga/effects";
-import { waitAndSubmitMessageToServer } from "../../api/sendPostToServer";
+import { waitAndSubmitMessageToServer } from "../../api/waitAndSubmitMessageToServer";
 import { MAX_BACKOFF_IN_MS, URI } from "../../config";
 import { pickMessageContent } from "../../utils/pickMessageContent";
 import { AwaitedReturnType } from "../../utils/ts/AwaitedReturnType";
