+++ src/state/state/IState.ts
@@ -1,6 +1,6 @@
 import { IAppState } from "./IAppState";
 import { INetworkState } from "./INetworkState";
-import { IMessagesState } from "./IPostsState";
+import { IMessagesState } from "./IMessagesState";
 import { ISubmitMessageState } from "./ISubmitMessageState";
 
 export interface IState {
